+++ src/models/index.ts
@@ -16,4 +16,3 @@ export * from './chat';
 export * from './toolkits';
 export * from './templates';
 export * from './nextSteps';
-export * from './actionButtons';